% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airzones.R
\docType{data}
\name{airzones}
\alias{airzones}
\title{British Columbia Air Zones}
\format{An \code{sf} polygons object with the boundaries of British Columbia's
seven air zones.

The data contains one column, \code{Airzone}, with the name of the Air Zone.}
\source{
Original data from the \href{https://catalogue.data.gov.bc.ca/dataset/e8eeefc4-2826-47bc-8430-85703d328516}{B.C. Data Catalogue},
under the \href{https://www2.gov.bc.ca/gov/content?id=A519A56BC2BF44E4A008B33FCF527F61}{Open Government Licence - British Columbia}.
}
\usage{
airzones
}
\description{
Boundaries of British Columbia Air Zones, which are areas of the province
for reporting on Canadian Ambient Air Quality Standards
(\href{http://www.ccme.ca/en/current_priorities/air/index.html}{CAAQS}), and
implementing the Air Quality Management System. For more information on Air
Zones in B.C. see \url{https://www2.gov.bc.ca/gov/content?id=8FA84ADC845A49AB93E525A58D9D4A97}.
}
\keyword{datasets}
