% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_cities.R
\docType{data}
\name{bc_cities}
\alias{bc_cities}
\title{BC Major Cities Points 1:2,000,000 (Digital Baseline Mapping)}
\format{An \code{sf} points object with the points for British Columbia's
major population centre.}
\source{
Original data from
\href{https://catalogue.data.gov.bc.ca/dataset/bc-major-cities-points-1-2-000-000-digital-baseline-mapping}{DataBC},
under the
\href{https://www2.gov.bc.ca/gov/content?id=A519A56BC2BF44E4A008B33FCF527F61}{Open
Government Licence - British Columbia}.
}
\usage{
bc_cities
}
\description{
Layer for BC's major population centres.
}
\keyword{datasets}
