% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bc_neighbours.R
\docType{data}
\name{bc_neighbours}
\alias{bc_neighbours}
\title{Boundary of British Columbia and provinces/states that border British Columbia}
\format{An object of class \code{sf} (inherits from \code{data.frame}) with 8 rows and 5 columns.}
\usage{
bc_neighbours
}
\description{
Boundary of British Columbia and provinces/states that border British Columbia bounded by a square projected box.
This layer is primarily intended for plotting purposes.
}
\details{
Downloaded from \href{https://www.naturalearthdata.com/}{Natural Earth} via
\href{https://github.com/ropensci/rnaturalearth}{rnaturalearth} and clipped to the area within a bounding box
then re-projected to BC Albers projection. Some attribute fields were removed, and other attribute
field names were changed.
}
\keyword{datasets}
