% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bec.R
\docType{data}
\name{bec}
\alias{bec}
\title{British Columbia BEC Map}
\format{An \code{sf} polygons object with B.C.'s Biogeoclimatic Ecosystem
Classification (BEC) Zone/Subzone/Variant/Phase map}
\source{
Original data from the
\href{https://catalogue.data.gov.bc.ca/dataset/f358a53b-ffde-4830-a325-a5a03ff672c3}{B.C. Data Catalogue},
under the
\href{https://www2.gov.bc.ca/gov/content?id=A519A56BC2BF44E4A008B33FCF527F61}{Open
Government Licence - British Columbia}.
}
\usage{
bec
}
\description{
The current and most detailed version of the approved corporate provincial
digital Biogeoclimatic Ecosystem Classification (BEC)
Zone/Subzone/Variant/Phase map (version 10, August 31st, 2016).
}
\keyword{datasets}
