% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoprovinces.R
\docType{data}
\name{ecoprovinces}
\alias{ecoprovinces}
\title{British Columbia Ecoprovinces}
\format{An \code{sf} polygons object with the boundaries of British Columbia's
ten ecoprovinces.}
\source{
Original data from the
\href{https://catalogue.data.gov.bc.ca/dataset/51832f47-efdf-4956-837a-45fc2c9032dd}{B.C. Data Catalogue},
under the
\href{https://www2.gov.bc.ca/gov/content?id=A519A56BC2BF44E4A008B33FCF527F61}{Open
Government Licence - British Columbia}.
}
\usage{
ecoprovinces
}
\description{
An ecoprovince is an area with consistent climatic or oceanography, relief
and regional landforms. There are one oceanic, two marine / terrestrial and seven
terrestrial ecoprovinces occurring within British Columbia. Ecoprovinces are
meant to be mapped at 1:2,000,000 for use in provincial state of the
environment reporting
}
\keyword{datasets}
