% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecosections.R
\docType{data}
\name{ecosections}
\alias{ecosections}
\title{British Columbia Ecosections}
\format{An \code{sf} polygons object with the boundaries of British Columbia's
114 ecosetions.}
\source{
Original data from the
\href{https://catalogue.data.gov.bc.ca/dataset/ccc01f43-860d-4583-8ba4-e72d8379441e}{B.C. Data Catalogue},
under the
\href{https://www2.gov.bc.ca/gov/content?id=A519A56BC2BF44E4A008B33FCF527F61}{Open
Government Licence - British Columbia}.
}
\usage{
ecosections
}
\description{
Ecosections are areas with minor physiographic and macroclimatic or
oceanographic variations. There are 114 ecosections in British Columbia
varying from pure marine units to pure terrestrial units. Ecosections are
meant to be mapped at small scales (1:250,000) for resource emphasis
and area planning
}
\keyword{datasets}
