% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr_areas.R
\docType{data}
\name{nr_areas}
\alias{nr_areas}
\title{British Columbia Natural Resource (NR) Areas}
\format{An \code{sf} polygons object with the boundaries of British Columbia's
3 Natural Resource (NR) Areas.}
\source{
Original data from
\href{https://catalogue.data.gov.bc.ca/dataset/natural-resource-nr-areas}{DataBC},
under the
\href{https://www2.gov.bc.ca/gov/content?id=A519A56BC2BF44E4A008B33FCF527F61}{Open
Government Licence - British Columbia}.
}
\usage{
nr_areas
}
\description{
British Columbia Natural Resource (NR) Areas are administrative boundaries
established by the Ministry of Forests, Lands and Natural Resource Operations.
These boundaries are an amalgamation of all NR Regions within the supplied boundaries.
}
\keyword{datasets}
