% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr_districts.R
\docType{data}
\name{nr_districts}
\alias{nr_districts}
\title{British Columbia Natural Resource (NR) Districts}
\format{An \code{sf} polygons object with the boundaries of British Columbia's
Natural Resource (NR) Districts.}
\source{
Original data from
\href{https://catalogue.data.gov.bc.ca/dataset/natural-resource-nr-district}{DataBC},
under the
\href{https://www2.gov.bc.ca/gov/content?id=A519A56BC2BF44E4A008B33FCF527F61}{Open
Government Licence - British Columbia}.
}
\usage{
nr_districts
}
\description{
British Columbia Natural Resource (NR) Districts are administrative boundaries
established by the Ministry of Forests, Lands and Natural Resource Operations.
}
\keyword{datasets}
