% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr_regions.R
\docType{data}
\name{nr_regions}
\alias{nr_regions}
\title{British Columbia Natural Resource (NR) Regions}
\format{An \code{sf} polygons object with the boundaries of British Columbia's
Natural Resource (NR) Regions, administrative boundaries established within NR Areas.}
\source{
Original data from \href{http://catalogue.data.gov.bc.ca/dataset/natural-resource-nr-regions}{DataBC},
under the \href{https://www2.gov.bc.ca/gov/content?id=A519A56BC2BF44E4A008B33FCF527F61}{Open Government Licence - British Columbia}.
}
\usage{
nr_regions
}
\description{
British Columbia Natural Resource (NR) Regions are administrative boundaries
established by the Ministry of Forests, Lands and Natural Resource Operations.
These boundaries are designated by the Lieutenant Governor in council and published
as regulations which establishes the Ministry's management areas.
}
\keyword{datasets}
