% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_districts.R
\docType{data}
\name{regional_districts}
\alias{regional_districts}
\title{British Columbia Ecoregions}
\format{An \code{sf} polygons object with the boundaries of British Columbia's
regional districts.}
\source{
Original data from
\href{https://catalogue.data.gov.bc.ca/pt_BR/dataset/d1aff64e-dbfe-45a6-af97-582b7f6418b9}{DataBC},
under the
\href{http://www.data.gov.bc.ca/local/dbc/docs/license/OGL-vbc2.0.pdf}{Open
Government License - British Columbia}.
}
\usage{
regional_districts
}
\description{
An ecoregion is an area with major physiographic and minor macroclimatic or
oceanographic variation. There are 43 ecoregion in British Columbia of which
39 are terrestrial. Ecodivsions are meant to be mapped at 1:500,000 for
regional strategic planning
}
\keyword{datasets}
