% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/municipalities.R
\docType{data}
\name{municipalities}
\alias{municipalities}
\title{British Columbia Municipalities}
\format{An \code{sf} polygons object with the boundaries of British Columbia's
municipalities.}
\source{
Original data from the
\href{https://catalogue.data.gov.bc.ca/dataset/e3c3c580-996a-4668-8bc5-6aa7c7dc4932}{B.C. Data Catalogue},
under the
\href{https://www2.gov.bc.ca/gov/content?id=A519A56BC2BF44E4A008B33FCF527F61}{Open
Government Licence - British Columbia}.
}
\usage{
municipalities
}
\description{
Legally defined Municipal polygons were drawn from metes and bounds descriptions
as written in Letters Patent for Municipalities in the province of British Columbia.
}
\keyword{datasets}
