% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regional_districts.R
\docType{data}
\name{regional_districts}
\alias{regional_districts}
\title{British Columbia Regional Districts}
\format{An \code{sf} polygons object with the boundaries of British Columbia's
regional districts.}
\source{
Original data from the
\href{https://catalogue.data.gov.bc.ca/pt_BR/dataset/d1aff64e-dbfe-45a6-af97-582b7f6418b9}{B.C. Data Catalogue},
under the
\href{https://www2.gov.bc.ca/gov/content?id=A519A56BC2BF44E4A008B33FCF527F61}{Open
Government Licence - British Columbia}.
}
\usage{
regional_districts
}
\description{
Legally defined Regional District polygons were drawn from metes and bounds descriptions
as written in Letters Patent for Regional Districts in the province of British Columbia.
}
\keyword{datasets}
